+++ src/cpp/session/include/session/worker_safe/session/SessionClientEvent.hpp
@@ -124,6 +124,7 @@ extern const int kInstallShiny;
 extern const int kSuspendAndRestart;
 extern const int kDataViewChanged;
 extern const int kViewFunction;
+extern const int kShowMarkers;
 }
    
 class ClientEvent