+++ src/gwt/src/org/rstudio/studio/client/server/remote/RemoteServer.java
@@ -3812,6 +3812,12 @@ public class RemoteServer implements Server
                   requestCallback);
    }
    
+   @Override
+   public void clearAllMarkers(ServerRequestCallback<Void> requestCallback)
+   {
+      sendRequest(RPC_SCOPE, "clear_all_markers", requestCallback);
+   }
+   
    private String clientId_;
    private double clientVersion_ = 0;
    private boolean listeningForEvents_;
@@ -4124,4 +4130,5 @@ public class RemoteServer implements Server
    private static final String GET_PACKRAT_STATUS = "get_packrat_status";
    private static final String PACKRAT_BOOTSTRAP = "packrat_bootstrap";
    private static final String GET_PENDING_ACTIONS = "get_pending_actions";
+  
 }