+++ NEWS.md
@@ -107,6 +107,7 @@
 
 * Prevent error dialog when getOption("repos") is an unnamed vector
 * Fix for regex Find/Replace lockup with empty strings 
+* Find in Files now always activates result pane